added login & middleware

+++ routes/web.php
@@ -19,31 +19,44 @@ use App\Http\Controllers\ListingController;
 // index
 Route::get('/', [ListingController::class, 'index']);
 // Create new Gig
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // Storing data to db
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 // Show route form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 
 
 // Update Data after edit 
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 
 // Delete
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 // User Routes
 // Register route
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // Store us
 Route::post('/users', [UserController::class, 'store']);
 
 // Login
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->middleware('guest')->name('login');
+
+// Login user
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
+
+// Logout
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+
+
+
+
+
+
 
 // Route Modal Binding -- Show Single -- Must be put at the bottom 
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
